# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Topping CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/score:
    post:
      operationId: scoreLead
      tags: [leads]
      summary: Calculate the AI score for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead with updated AI score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/log-activity:
    post:
      operationId: logLeadActivity
      tags: [leads]
      summary: Log an activity (email, call, meeting, etc.) against a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogActivityInput"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/schedule-meeting:
    post:
      operationId: scheduleLeadMeeting
      tags: [leads]
      summary: Schedule a meeting with a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleMeetingInput"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /leads/marketing-report:
    post:
      operationId: generateMarketingReport
      tags: [leads]
      summary: Generate an AI marketing performance report (optionally email it)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketingReportInput"
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketingReport"

  # ─── DEALS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, name, stage]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]
        stage:
          type: string
          enum: [new, contacted, qualified, proposal, negotiation, closed_won, closed_lost]
        source:
          type: ["string", "null"]
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        status:
          type: string
          enum: [new_lead, contacted_email, contacted_call, contacted_linkedin, meeting_scheduled, meeting_done, proposal_sent, negotiating, closed_won, closed_lost, not_interested, nurturing]
        aiScore:
          type: ["integer", "null"]
        scoreReason:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
          enum: [HOT, WARM, COLD, null]
        aiNextAction:
          type: ["string", "null"]
        gtaNode:
          type: ["integer", "null"]
        activityType:
          type: ["string", "null"]
          enum: [email, call, meeting, linkedin_message, follow_up, null]
        nextActionDate:
          type: ["string", "null"]
        meetingDate:
          type: ["string", "null"]
        emailsSent:
          type: integer
        emailsReceived:
          type: integer
        createdAt:
          type: string
    LeadInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        companyId:
          type: integer
        stage:
          type: string
        source:
          type: string
        ownerId:
          type: integer
        value:
          type: number
        notes:
          type: string
        linkedinUrl:
          type: string
        industry:
          type: string
        status:
          type: string
        gtaNode:
          type: integer
        activityType:
          type: string
        nextActionDate:
          type: string
        meetingDate:
          type: string
        emailsSent:
          type: integer
        emailsReceived:
          type: integer
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        companyId:
          type: integer
        stage:
          type: string
        source:
          type: string
        ownerId:
          type: integer
        value:
          type: number
        notes:
          type: string
        linkedinUrl:
          type: string
        industry:
          type: string
        status:
          type: string
        gtaNode:
          type: integer
        activityType:
          type: string
        nextActionDate:
          type: string
        meetingDate:
          type: string
        emailsSent:
          type: integer
        emailsReceived:
          type: integer
    LogActivityInput:
      type: object
      required: [activityType]
      properties:
        activityType:
          type: string
          enum: [email, call, meeting, linkedin_message, follow_up]
        notes:
          type: string
        status:
          type: string
        nextActionDate:
          type: string
    ScheduleMeetingInput:
      type: object
      required: [startTime, endTime]
      properties:
        title:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        location:
          type: string
        onlineLink:
          type: string
        notes:
          type: string
...
    MarketingReportInput:
      type: object
      properties:
        email:
          type: string
    MarketingReport:
      type: object
      required: [report]
      properties:
        report:
          type: string
        emailed:
          type: boolean

    # ─── DEAL ─────────────────────────────────────────────────────────────────
...